# contentstudio-social-analytics-go-main/docs/openapi.yaml
openapi: "3.0.3"
info:
  title: ContentStudio Social Analytics API
  version: "1.0.0"
  description: |
    Analytics data API for ContentStudio social media insights.
    Serves ClickHouse-backed analytics for Facebook, Instagram, LinkedIn, YouTube,
    TikTok, Pinterest, Twitter, GMB, Meta Ads, and cross-platform Overview V2.

    **Authentication** — every request (except `/health`) must carry one of:
    - `Authorization: Bearer <jwt>` — workspace JWT
    - `X-API-Key: <key>` — API key
    - `X-Shareable-ID: <id>` — shareable report link (analytics paths only)
servers:
  - url: http://localhost:8080
    description: Local development
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []

# ─────────────────────────────────────────────────────────────
# PATHS
# ─────────────────────────────────────────────────────────────
paths:
...
  /api/v1/immediate-work:
    post:
      tags: [Work Orders]
      summary: Enqueue an immediate analytics fetch for one account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImmediateWorkRequest"
      responses:
        "200":
          description: Work order accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/competitor-work:
    post:
      tags: [Work Orders]
      summary: Enqueue a competitor analytics fetch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompetitorWorkRequest"
      responses:
        "200":
          description: Work order accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/listening-work:
    post:
      tags: [Work Orders]
      summary: Enqueue a listening data fetch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Work order accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"

  # ─────────────────────────────────────────────────────────
  # LINKEDIN
  # ─────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Workspace JWT — Authorization: Bearer <token>"
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Static API key
    ShareableAuth:
      type: apiKey
      in: header
      name: X-Shareable-ID
      description: Shareable report link ID (analytics paths only)
...
  schemas:
...
    StatusResponse:
      type: object
      properties:
        status:
          type: boolean
...
    ImmediateWorkRequest:
      type: object
      properties:
        workspace_id:
          type: string
        platform:
          type: string
        account_id:
          type: string
    CompetitorWorkRequest:
      type: object
      properties:
        workspace_id:
          type: string
        report_id:
          type: string
        platform:
          type: string

    # ── LinkedIn ────────────────────────────────────────────
...